Uses Arduino's avr-objcopy on system that does not ship a native one
Some systems (like OS X) does not ship a native objcopy utility, hence
CMAKE_OBJCOPY is empty arduino-cmake rules need it to be set. However
arduino's SDK does ship one. In case the system does not have one,
this patch will look for the arduino's one and ensure that
CMAKE_OBJCOPY is set to that one for the project.