Fixing problem when using HDRS, LIBS, AFLAGS options with generate_avr_firmware
These options might be set or left out (i.e. empty).
So the variables holding the values (e.g. INPUT_HDRS for HDRS)
are checked whether they are empty or not. If they are not
empty the according option name should be added to the front
of the list.

This has been done by setting the variable with a string
containing the option name followed by a space and the
variable content.
The result is not a list with the option name as first
element. Since the space character doesn't work as a
list item delimiter, the first element is a mixture
of option name, space and the old first element in the list.
This didn't get parsed later and resulted in
all these options getting added to the SRCS option.

I changed the update of the lists to use the cmake
list functions. The proper option name gets now inserted into
the lists at the beginning.